Moved session tracking logic to its own component and also changed the way we track the session, so we don't trigger so many api calls
